Move GameSession class into dedicated header file
