+++ server/GameServer.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "GameSession.h"
 #include "Packet.h"
 #include <ctime>
 #include <cstring>
@@ -59,13 +60,6 @@ public:
 	void WaitForEvent();
 
 private:
-	enum GameServerGameFlags
-	{
-		GameSessionExpected = 1,
-		GameSessionReceived = 2,
-		GameSessionUpdateRetrySent = 4,
-	};
-
 	enum GameServerPacketErrorCode
 	{
 		// No packet
@@ -75,37 +69,6 @@ private:
 		PacketSizeFieldBad = -3,
 		PacketTypeBad = -4,
 		PacketChecksumBad = -5,
-	};
-
-	struct GameSession
-	{
-		GUID sessionIdentifier;
-		sockaddr_in addr;
-		std::time_t time;
-		unsigned int clientRandValue;
-		unsigned int serverRandValue;
-		unsigned int flags = 0;
-		CreateGameInfo createGameInfo;
-
-		inline bool SocketAddressMatches(const sockaddr_in& socketAddress) const
-		{
-			return memcmp(&addr, &socketAddress, sizeof(socketAddress)) == 0;
-		}
-
-		inline bool IsGameSessionExpected() const
-		{
-			return (flags & GameServerGameFlags::GameSessionExpected) != 0;
-		}
-
-		inline bool IsGameSessionReceived() const
-		{
-			return (flags & GameServerGameFlags::GameSessionReceived) != 0;
-		}
-
-		inline bool IsGameSessionUpdateRetrySent() const
-		{
-			return (flags & GameServerGameFlags::GameSessionUpdateRetrySent) != 0;
-		}
 	};
 
 	void AllocSocket(SOCKET& socket, unsigned short port);
